+++ src/Formats/OBJ.cpp
@@ -41,7 +41,7 @@ namespace
 
     void WriteFaces(std::ostream& o, osc::Mesh const& mesh, osc::ObjWriterFlags flags)
     {
-        if (mesh.getTopography() != osc::MeshTopography::Triangles)
+        if (mesh.getTopology() != osc::MeshTopology::Triangles)
         {
             return;
         }